apps/api/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic-settings==2.2.1
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pypdf==4.2.0
python-docx==1.1.2
pandas==2.2.2
openpyxl==3.1.3
openai==1.30.1
pgvector==0.2.5
httpx==0.27.0
pytest==8.2.2
